использование общего app_bar для групп администратора
